Rename `session.rs` to `ssl.rs`
